moved new functions to `version` to keep api_version a pure static file
