+++ InvenTree/InvenTree/version.py
@@ -16,7 +16,7 @@ from django.conf import settings
 
 from dulwich.repo import NotGitRepository, Repo
 
-from .api_version import INVENTREE_API_TEXT_DATA, INVENTREE_API_VERSION
+from .api_version import INVENTREE_API_TEXT, INVENTREE_API_VERSION
 
 # InvenTree software version
 INVENTREE_SW_VERSION = "0.13.0 dev"
@@ -140,6 +140,33 @@ def isInvenTreeUpToDate():
 def inventreeApiVersion():
     """Returns current API version of InvenTree."""
     return INVENTREE_API_VERSION
+
+
+def parse_version_text():
+    """Parse the version text to structured data."""
+    patched_data = INVENTREE_API_TEXT.split("\n\n")
+    # Remove first newline on latest version
+    patched_data[0] = patched_data[0].replace("\n", "", 1)
+
+    version_data = {}
+    for version in patched_data:
+        data = version.split("\n")
+
+        version_split = data[0].split(' -> ')
+        version_detail = version_split[1].split(':', 1) if len(version_split) > 1 else ['', ]
+        new_data = {
+            "version": version_split[0].strip(),
+            "date": version_detail[0].strip(),
+            "gh": version_detail[1].strip() if len(version_detail) > 1 else None,
+            "text": data[1:],
+            "latest": False,
+        }
+        version_data[new_data["version"]] = new_data
+    return version_data
+
+
+INVENTREE_API_TEXT_DATA = parse_version_text()
+"""Pre-processed API version text."""
 
 
 def inventreeApiText(versions: int = 10, start_version: int = 0):
